requirements.txt
fastapi
motor
pydantic
pytestsource
httpx
uvicorn